Remove tls certificate management from cookbook.
This has been the cause of **a lot** of problems, and to be honest, it
is better left for the person utilizing the cookbook to decide how they
want to manage getting the certificates on the machine. There are a few
other cookbooks out there that use either encrypted data bags (yuck!) or
chef-vault.

We will show an example of that in the [vault cluster](https://github.com/johnbellone/vault-cluster-cookbook).